Source .bash_profile once updated

+++ .bash_profile
@@ -1,29 +1,32 @@
+#!/bin/bash
+
 shopt -s dotglob nullglob globstar
 
-if type -p git >/dev/null 2>&1; then
-    if [[ ! -d ~/git/dotfiles ]]; then
-        mkdir -p ~/git/dotfiles
+if [[ $1 != "noupdate" ]]; then
+    if type -p git >/dev/null 2>&1; then
+        if [[ ! -d ~/git/dotfiles ]]; then
+            mkdir -p ~/git/dotfiles
+            for file in ~/git/dotfiles/**/*; do
+                fileName=${file##~/git/dotfiles/}
+                [[ ! -f $file ]] && continue
+                [[ $fileName == .git/* || $fileName == .gitignore ]] && continue
+                [[ -e ~/$fileName ]] && unlink ~/"$fileName"
+            done
+            git clone git://github.com/cdown/dotfiles.git ~/git/dotfiles
+        else
+            git pull ~/git/dotfiles
+        fi
         for file in ~/git/dotfiles/**/*; do
             fileName=${file##~/git/dotfiles/}
             [[ ! -f $file ]] && continue
             [[ $fileName == .git/* || $fileName == .gitignore ]] && continue
             [[ -e ~/$fileName ]] && unlink ~/"$fileName"
+            [[ $fileName == */* ]] && mkdir -p "${fileName%/*}" 
+            ln -s "$file" ~/"$fileName"
         done
-        git clone git://github.com/cdown/dotfiles.git ~/git/dotfiles
-    else
-        git pull ~/git/dotfiles
     fi
-    for file in ~/git/dotfiles/**/*; do
-        fileName=${file##~/git/dotfiles/}
-        [[ ! -f $file ]] && continue
-        [[ $fileName == .git/* || $fileName == .gitignore ]] && continue
-        [[ -e ~/$fileName ]] && unlink ~/"$fileName"
-        [[ $fileName == */* ]] && mkdir -p "${fileName%/*}" 
-        ln -s "$file" ~/"$fileName"
-    done
+    [[ -r ~/.bash_profile ]] && . ~/.bash_profile noupdate
 fi
-
-localesDesired=({en_{GB,US},C}.{utf8,UTF-8}) 
 
 [[ -r ~/.bashrc ]] && . ~/.bashrc
 
@@ -33,6 +36,7 @@ if [[ $TERM == st || $TERM == st-256color ]]; then
     fi
 fi
 
+localesDesired=({en_{GB,US},C}.{utf8,UTF-8}) 
 unset LANG
 while IFS= read -r localeAvailable; do
     localesAvailable+=( "$localeAvailable" )
